outgoing_links belongs in helpers, not component
